% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rh.R
\name{add_rh}
\alias{add_rh}
\title{Add relative humidity to APSIM weather data}
\usage{
add_rh(apsim, standard, output = NULL, overwrite = FALSE)
}
\arguments{
\item{apsim}{A character string specifying the path to the APSIM weather file.}

\item{standard}{A character string specifying the path to the standard weather file containing relative humidity data.}

\item{output}{A character string specifying the path to save the modified APSIM weather file with relative humidity data. If NULL, the function will not save the file.}

\item{overwrite}{A logical value indicating whether to overwrite the output file if it already exists. Default is FALSE.}
}
\value{
A data frame containing the new weather records with relative humidity added.
}
\description{
This function adds relative humidity data to APSIM weather records based on a standard weather file.
}
