% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{addMarkers}
\alias{addMarkers}
\title{Add marker for each site}
\usage{
addMarkers(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Add marker for each site
}
