% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerRes.R
\name{addResult,result,result-method}
\alias{addResult,result,result-method}
\title{Add new result name and type}
\usage{
\S4method{addResult}{result,result}(object, an.res)
}
\arguments{
\item{object}{A result object.}

\item{an.res}{Another result object.}
}
\description{
Add new result name and type
}
