% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerRes.R
\name{addResult}
\alias{addResult}
\title{Add new result name and type}
\usage{
addResult(object, an.res)
}
\arguments{
\item{object}{A result object.}

\item{...}{Not used}
}
\description{
Add new result name and type
}
