% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeWeatherRecords.R
\docType{methods}
\name{changeWeatherRecords}
\alias{changeWeatherRecords}
\alias{changeWeatherRecords,WeaAna-method}
\alias{changeWeatherRecords,WeaAna,WeaAna-method}
\title{Change weather records}
\usage{
changeWeatherRecords(object, ...)

\S4method{changeWeatherRecords}{WeaAna}(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{New weather records}
}
\value{
A new WeaAna object with updated records
}
\description{
Change weather records

Change weather records
}
