% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeWeatherRecords.R
\docType{methods}
\name{changeWeatherRecords,WeaAna-method}
\alias{changeWeatherRecords,WeaAna-method}
\title{Change weather records}
\usage{
\S4method{changeWeatherRecords}{WeaAna}(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{New weather records}
}
\description{
Change weather records
}
