% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeWeatherRecords.R
\name{changeWeatherRecords}
\alias{changeWeatherRecords}
\title{Change weather records}
\usage{
changeWeatherRecords(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Change weather records
}
