% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2Records.R
\name{convert2Records}
\alias{convert2Records}
\title{Convert a data frame to weaana class}
\usage{
convert2Records(infor, records)
}
\arguments{
\item{infor}{A list or data frame of site information}

\item{records}{A data frame will convert to records}
}
\description{
Convert a data frame to weaana class
}
