% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWeatherRecords.r
\name{createWeaAna}
\alias{createWeaAna}
\title{create WeaAna class}
\usage{
createWeaAna(mets)
}
\arguments{
\item{mets}{A list contained information of weather records.}
}
\description{
create WeaAna class
}
