% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{delMarkers,WeaAna-method}
\alias{delMarkers,WeaAna-method}
\title{Delete marker for each site}
\usage{
\S4method{delMarkers}{WeaAna}(object, name = NULL)
}
\arguments{
\item{object}{A WeaAna object.}

\item{name}{A vector of name to remove}
}
\description{
Delete marker for each site
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
addMarkers( records, model = c( "SILO", "A2" ) ) 
delMarkers( records, name = "model" )

}
