% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{delMarkers}
\alias{delMarkers}
\title{Delete marker for each site}
\usage{
delMarkers(object, name)
}
\arguments{
\item{object}{A WeaAna object.}

\item{name}{A vector of name to remove}
}
\description{
Delete marker for each site
}
