% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{DOY}
\alias{DOY}
\title{Calculate day of year according to date}
\usage{
DOY(Date, format = NULL)
}
\arguments{
\item{Date}{a string or date vector.}

\item{format}{date format when Date is a string vector}
}
\description{
Calculate day of year according to date
}
