% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{findResType,WeaAna,character-method}
\alias{findResType,WeaAna,character-method}
\title{Get result type by name}
\usage{
\S4method{findResType}{WeaAna,character}(object, name)
}
\arguments{
\item{object}{A WeaAna object.}

\item{name}{Result name.}
}
\description{
Get result type by name
}
