% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{findResType}
\alias{findResType}
\title{Get result type by name}
\usage{
findResType(object, name)
}
\arguments{
\item{object}{A WeaAna object.}

\item{name}{Result name.}

\item{...}{Not used}
}
\description{
Get result type by name
}
