% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFrostDay.R
\docType{methods}
\name{firstFrostDay,numeric-method}
\alias{firstFrostDay,numeric-method}
\title{Calculate the last frost}
\usage{
\S4method{firstFrostDay}{numeric}(
  object,
  stress = waGetPara("extreme$mint$frost.night$value"),
  latitude = -20
)
}
\arguments{
\item{object}{A numeric object.}

\item{stress}{The stress temperature for frost}

\item{latitude}{Latitude of site.}
}
\description{
Calculate the last frost
}
