% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFrostDay.R
\docType{methods}
\name{firstFrostDay,WeaAna-method}
\alias{firstFrostDay,WeaAna-method}
\title{Calculate the last frost}
\usage{
\S4method{firstFrostDay}{WeaAna}(object,
  stress = waGetPara("extreme$mint$frost.night$value"),
  yrange = waGetPara("yrange"))
}
\arguments{
\item{object}{A WeaAna object.}

\item{stress}{The stress temperature for frost}

\item{yrange}{The year range for statistics}
}
\description{
Calculate the last frost
}
