% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstFrostDay.R
\name{firstFrostDay}
\alias{firstFrostDay}
\title{Calculate the first frost event}
\usage{
firstFrostDay(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate the first frost event
}
