% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstHeatDay.R
\docType{methods}
\name{firstHeatDay,numeric-method}
\alias{firstHeatDay,numeric-method}
\title{Calculate the first heat day}
\usage{
\S4method{firstHeatDay}{numeric}(object,
  stress = waGetPara("extreme$maxt$hot.day$value"), latitude = -20,
  conse = 1, offset = 0)
}
\arguments{
\item{object}{A numeric object.}

\item{stress}{The stress temperature for frost}

\item{latitude}{Latitude of site.}

\item{conse}{Number of consecutive days}

\item{offset}{offset}
}
\description{
Calculate the first heat day
}
