% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstHeatDay.R
\docType{methods}
\name{firstHeatDay,WeaAna-method}
\alias{firstHeatDay,WeaAna-method}
\title{Calculate the first heat day}
\usage{
\S4method{firstHeatDay}{WeaAna}(
  object,
  stress = waGetPara("extreme$maxt$hot.day$value"),
  yrange = waGetPara("yrange"),
  conse = 1,
  offset = 0
)
}
\arguments{
\item{object}{A WeaAna object.}

\item{stress}{The stress temperature for frost}

\item{yrange}{The year range for statistics}

\item{conse}{Number of consecutive days}

\item{offset}{offset
The first heat day of each year will return if prob equal to NULL.}
}
\description{
Calculate the first heat day
}
