% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstHeatDay.R
\name{firstHeatDay}
\alias{firstHeatDay}
\title{Calculate the first heat day}
\usage{
firstHeatDay(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate the first heat day
}
