% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{getAvaiVars,WeaAna-method}
\alias{getAvaiVars,WeaAna-method}
\title{Get available variables of weather records}
\usage{
\S4method{getAvaiVars}{WeaAna}(object)
}
\arguments{
\item{object}{A WeaAna object.}
}
\description{
Get available variables of weather records
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
getAvaiVars( records )
}
