% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getAvaiVars,WeaAnaSite-method}
\alias{getAvaiVars,WeaAnaSite-method}
\title{Get available variables of weather records}
\usage{
\S4method{getAvaiVars}{WeaAnaSite}(object)
}
\arguments{
\item{object}{A WeaAnaSite object.}
}
\description{
Get available variables of weather records
}
