% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getAvaiVars}
\alias{getAvaiVars}
\title{Get available variables of weather records}
\usage{
getAvaiVars(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Get available variables of weather records
}
