% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{getFilenames,WeaAna-method}
\alias{getFilenames,WeaAna-method}
\title{Get filename}
\usage{
\S4method{getFilenames}{WeaAna}(object)
}
\arguments{
\item{object}{A WeaAna object.}
}
\description{
Get filename
}
