% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{getResults}
\alias{getResults}
\alias{getResults,WeaAna-method}
\alias{getResults,WeaAna,WeaAna-method}
\title{Get results by name}
\usage{
getResults(object, ...)

\S4method{getResults}{WeaAna}(object, name = NULL)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}

\item{name}{The results with these name are returned.}
}
\value{
Results for weaana object
}
\description{
Get results by name

Get results by name
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
getResults( records, "result" )
}
