% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{getStationInforByString,WeaAnaSite-method}
\alias{getStationInforByString,WeaAnaSite-method}
\title{Get a string which contains the basic information of weather station.}
\usage{
\S4method{getStationInforByString}{WeaAnaSite}(object, yrange = NULL)
}
\arguments{
\item{object}{A WeaAna object.}

\item{yrange}{Year range.}
}
\description{
Get a string which contains the basic information of weather station.
}
