% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getStationInforByString}
\alias{getStationInforByString}
\title{Get a string which contains the basic information of weather station.}
\usage{
getStationInforByString(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Get a string which contains the basic information of weather station.
}
