% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{[,WeaAna-method}
\alias{[,WeaAna-method}
\title{Getter to access the weather data at a specific position.}
\usage{
\S4method{[}{WeaAna}(x, i, j, drop)
}
\arguments{
\item{x}{A WeaAna object.}

\item{i}{the specific position which will access.}

\item{j}{None use parameter.}

\item{drop}{None use parameter.}
}
\value{
A WeaAnaSite object at the position i.
}
\description{
Getter to access the weather data at a specific position.
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
records[1]
records[1:2]
records[2:2]

}
