% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{getWeaAnaSiteByPos,WeaAna-method}
\alias{getWeaAnaSiteByPos,WeaAna-method}
\title{Plot site position in a map}
\usage{
\S4method{getWeaAnaSiteByPos}{WeaAna}(object, pos)
}
\arguments{
\item{object}{A WeaAna object.}

\item{pos}{The position would be return.}
}
\description{
Plot site position in a map
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
getWeaAnaSiteByPos( records, 1 )
}
