% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getWeaAnaSiteByPos}
\alias{getWeaAnaSiteByPos}
\title{Get WeaAnaSite object by position}
\usage{
getWeaAnaSiteByPos(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Get WeaAnaSite object by position
}
