% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{getWeatherRecords}
\alias{getWeatherRecords}
\alias{getWeatherRecords,WeaAna-method}
\alias{getWeatherRecords,WeaAna,WeaAna-method}
\title{Get all weather records by year range}
\usage{
getWeatherRecords(object, ...)

\S4method{getWeatherRecords}{WeaAna}(object, yrange = NULL, vars = "all", ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Other arguments}

\item{yrange}{Year range.}

\item{vars}{Variable}
}
\value{
A data frame with all weather records
}
\description{
Get all weather records by year range

Get all weather records by year range
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
getWeatherRecords( records, yrange = c( 2008, 2009 ) )
getWeatherRecords( records, yrange = c( 2008, 2009 ), length = 10 )

}
