% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getWeatherRecords,WeaAna-method}
\alias{getWeatherRecords,WeaAna-method}
\title{Get all weather records by year range}
\usage{
\S4method{getWeatherRecords}{WeaAna}(object, yrange = NULL, vars = "all", ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{yrange}{Year range.}

\item{vars}{Variable}

\item{...}{Other arguments}
}
\description{
Get all weather records by year range
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
getWeatherRecords( records, yrange = c( 2008, 2009 ) )
getWeatherRecords( records, yrange = c( 2008, 2009 ), length = 10 )

}
