% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{getWeatherRecords}
\alias{getWeatherRecords}
\title{Get all weather records by year range}
\usage{
getWeatherRecords(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Get all weather records by year range
}
