% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly2Daily.R
\name{hourly2Daily}
\alias{hourly2Daily}
\title{Convert hourly weather records for daily APSIM format}
\usage{
hourly2Daily(records)
}
\arguments{
\item{records}{hourly weather records}
}
\description{
Convert hourly weather records for daily APSIM format
}
