% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastFrostDay.R
\docType{methods}
\name{lastFrostDay,numeric-method}
\alias{lastFrostDay,numeric-method}
\title{Calculate the last frost}
\usage{
\S4method{lastFrostDay}{numeric}(
  object,
  stress = waGetPara("extreme$mint$frost.night$value"),
  latitude = -20,
  offset = 0
)
}
\arguments{
\item{object}{A numeric object.}

\item{stress}{The stress temperature for frost}

\item{latitude}{Latitude of site.}

\item{offset}{offset.}
}
\description{
Calculate the last frost
}
