% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastFrostDay.R
\docType{methods}
\name{lastFrostDay,WeaAna-method}
\alias{lastFrostDay,WeaAna-method}
\title{Calculate the last frost}
\usage{
\S4method{lastFrostDay}{WeaAna}(object,
  stress = waGetPara("extreme$mint$frost.night$value"),
  yrange = waGetPara("yrange"), offset = 0)
}
\arguments{
\item{object}{A WeaAna object.}

\item{stress}{The stress temperature for frost}

\item{yrange}{The year range for statistics}

\item{offset}{offset}
}
\description{
Calculate the last frost
}
