% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastFrostDay.R
\name{lastFrostDay}
\alias{lastFrostDay}
\title{Calculate the last frost}
\usage{
lastFrostDay(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate the last frost
}
