% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapYear.r
\name{leapYear}
\alias{leapYear}
\title{Judge a leap or common year}
\usage{
leapYear(year)
}
\arguments{
\item{year}{A numeric of year}
}
\value{
True for leap year or False for common year
}
\description{
Judge a leap or common year
}
