% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movExtreme.r
\name{maximumDays}
\alias{maximumDays}
\title{Calculate the maximum days which degree days is more than "key.degree.day"}
\usage{
maximumDays(degree.day, key.degree.day = 400)
}
\arguments{
\item{degree.day}{A vector of accumulated temperature}

\item{key.degree.day}{The periods for moving window}
}
\value{
The probability occuring a certain days extreme temperature
}
\description{
Calculate the maximum days which degree days is more than "key.degree.day"
}
