% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2RData.R
\name{Met2RData}
\alias{Met2RData}
\title{Convert met file to RData files}
\usage{
Met2RData(
  dataFiles = NULL,
  dataFolders = NULL,
  outputFolder = NULL,
  dataFormat = "APSIM"
)
}
\arguments{
\item{dataFiles}{A character vector to specify the path of weather data files.}

\item{dataFolders}{A character vector to specify the path of weather data folders.}

\item{outputFolder}{All RData files would be put into this folder. The default position}

\item{dataFormat}{The format of weather data file. 
is the original folder}
}
\description{
Convert met file to RData files
}
