% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.r
\name{mov.avg}
\alias{mov.avg}
\title{Use Calculate the moving average. For compatibility only.}
\usage{
mov.avg(x, k = 10, shift = "centre")
}
\arguments{
\item{x}{A vector to calculate moving average}

\item{k}{The moving windows}

\item{shift}{if shift = "centre", then values are shifted to centre. 
if shift = "begin", then values are at begin of period.
if shift = "end", then values are at end of period. 
The default value (centre) will be used if shift is other value.}
}
\value{
The moving average of vector x at moving windows n
}
\description{
Note that for n = odd, can average at central period. If n = even, must average at
end of period and then shift values
}
