% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.r
\name{mov.min}
\alias{mov.min}
\title{Calculate the moving minimum. For compatibility only.}
\usage{
mov.min(x, k, shift = "centre")
}
\arguments{
\item{x}{A vector to calculate moving minimum}

\item{k}{The moving windows}

\item{shift}{if shift = "centre", then values are shifted to centre.
if shift = "begin", then values are at begin of period.
if shift = "end", then values are at end of period.
The default value (centre) will be used if shift is other value.}
}
\value{
The moving minimum of vector x at moving windows k
}
\description{
Calculate the moving minimum. For compatibility only.
}
