% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.r
\name{mov}
\alias{mov}
\title{Calculate the moving values}
\usage{
mov(x, k = 10, shift = "centre", fun = "mean")
}
\arguments{
\item{x}{A vector to calculate moving values}

\item{k}{The moving windows}

\item{shift}{if shift = "centre", then values are shifted to centre. 
if shift = "begin", then values are at begin of period.
if shift = "end", then values are at end of period.
The default value (centre) will be used if shift is other value.}

\item{fun}{The method to calculate moving values. Curruntly, only
"mean", "max", "min", and "sum" are supported. A NULL will be returned
for any other values}
}
\value{
The moving value of vector x at moving windows k. A NULL will be returned
for any unsupported fun
}
\description{
Calculate the moving values
}
