% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mov.r
\name{mov.sum}
\alias{mov.sum}
\title{Calculate the moving sum. For compatibility only.}
\usage{
mov.sum(x, k, shift = "centre")
}
\arguments{
\item{x}{A vector to calculate moving sum}

\item{k}{The moving windows}

\item{shift}{if shift = "centre", then values are shifted to centre. 
if shift = "begin", then values are at begin of period.
if shift = "end", then values are at end of period.
The default value (centre) will be used if shift is other value.}
}
\value{
The moving sum of vector x at moving windows k
}
\description{
Calculate the moving sum. For compatibility only.
}
