% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movExtreme.r
\docType{methods}
\name{movExtremeByDay,WeaAna-method}
\alias{movExtremeByDay,WeaAna-method}
\title{Calculate the moving extreme temperature by days}
\usage{
\S4method{movExtremeByDay}{WeaAna}(
  object,
  indices = waGetPara("extreme"),
  period = waGetPara("mov.window"),
  shift = waGetPara("shift"),
  yrange = waGetPara("yrange"),
  numdays = waGetPara("numdays"),
  res.name = "movExtremeByDay"
)
}
\arguments{
\item{object}{A WeaAna object.}

\item{indices}{The data will be classify according the classes}

\item{period}{The periods for moving window}

\item{shift}{The method for calculating percentage of moving extreme temperature.}

\item{yrange}{The year range for statistics}

\item{numdays}{The maximum days for statistics}

\item{res.name}{The name for result}
}
\value{
The probability occurring a certain days extreme temperature
}
\description{
Calculate the moving extreme temperature by days
}
