% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movExtreme.r
\docType{methods}
\name{movExtremeByDegreeDay,WeaAna-method}
\alias{movExtremeByDegreeDay,WeaAna-method}
\title{Calculate the moving extreme temperature by commulated temperature}
\usage{
\S4method{movExtremeByDegreeDay}{WeaAna}(
  object,
  indices = waGetPara("extreme"),
  period = 400,
  shift = waGetPara("shift"),
  yrange = waGetPara("yrange"),
  numdays = waGetPara("numdays"),
  res.name = "movExtremeByDegreeDay",
  ...
)
}
\arguments{
\item{indices}{The data will be classify according the classes}

\item{period}{The periods for moving window}

\item{shift}{The method for calculating percentage of moving extreme temperature.}

\item{yrange}{The year range for statistics}

\item{numdays}{The maximum days for statistics}

\item{...}{Other argument to calculate thermal time}

\item{x}{The list which include years, days and data to calculate}
}
\description{
Calculate the moving extreme temperature by commulated temperature
}
