% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movExtreme.r
\name{movExtremeByDegreeDay}
\alias{movExtremeByDegreeDay}
\title{Calculate the moving extreme temperature by commulated temperature}
\usage{
movExtremeByDegreeDay(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate the moving extreme temperature by commulated temperature
}
