% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFrostDay.R
\docType{methods}
\name{numberFrostDay,numeric-method}
\alias{numberFrostDay,numeric-method}
\title{Calculate number of frost day}
\usage{
\S4method{numberFrostDay}{numeric}(object,
  stress = waGetPara("extreme$mint$frost.night$value"))
}
\arguments{
\item{object}{A numeric object.}

\item{stress}{The stress temperature for frost}
}
\description{
Calculate number of frost day
}
