% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFrostDay.R
\docType{methods}
\name{numberFrostDay,WeaAna-method}
\alias{numberFrostDay,WeaAna-method}
\title{Calculate number of frost day}
\usage{
\S4method{numberFrostDay}{WeaAna}(
  object,
  stress = waGetPara("extreme$mint$frost.night$value"),
  yrange = waGetPara("yrange")
)
}
\arguments{
\item{object}{A WeaAna object.}

\item{stress}{The stress temperature for frost}

\item{yrange}{The year range for statistics}
}
\description{
Calculate number of frost day
}
