% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFrostDay.R
\name{numberFrostDay}
\alias{numberFrostDay}
\title{Calculate number of frost day}
\usage{
numberFrostDay(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate number of frost day
}
