% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberHeatDay.R
\docType{methods}
\name{numberHeatDay,numeric-method}
\alias{numberHeatDay,numeric-method}
\title{Calculate number of heat day}
\usage{
\S4method{numberHeatDay}{numeric}(object, stress = waGetPara("extreme$maxt$hot.day$value"))
}
\arguments{
\item{object}{A numeric object.}

\item{stress}{The stress temperature for heat}
}
\description{
Calculate number of heat day
}
