% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberHeatDay.R
\docType{methods}
\name{numberHeatDay,WeaAna-method}
\alias{numberHeatDay,WeaAna-method}
\title{Calculate number of heat day}
\usage{
\S4method{numberHeatDay}{WeaAna}(
  object,
  stress = waGetPara("extreme$maxt$hot.day$value"),
  yrange = waGetPara("yrange")
)
}
\arguments{
\item{object}{A WeaAna object.}

\item{stress}{The stress temperature for heat}

\item{yrange}{The year range for statistics}
}
\description{
Calculate number of heat day
}
