% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberHeatDay.R
\name{numberHeatDay}
\alias{numberHeatDay}
\title{Calculate number of heat day}
\usage{
numberHeatDay(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate number of heat day
}
