% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGraphics.r
\name{makeOzViewports}
\alias{makeOzViewports}
\alias{makeOzLines}
\alias{ozGrob}
\alias{grid.ozGrob}
\title{All codes copy from library RGraphics. The xscale and yscale were changed to draw more figures in the right and top}
\usage{
makeOzViewports(ozRegion, xlim = ozRegion$rangex, ylim = ozRegion$rangey)

makeOzLines(ozRegion)

ozGrob(
  ozRegion,
  name = NULL,
  gp = NULL,
  vp = NULL,
  xlim = ozRegion$rangex,
  ylim = ozRegion$rangey
)

grid.ozGrob(...)
}
\arguments{
\item{ozRegion}{an object of class "ozRegion" for library ozRegion}

\item{name}{A character identifier for the grob}

\item{gp}{A gpar object}

\item{vp}{A viewport object}

\item{...}{For graphics parameters}
}
\description{
All codes copy from library RGraphics. The xscale and yscale were changed to draw more figures in the right and top

All codes copy from library RGraphics.

All codes copy from library RGraphics.

All codes copy from library RGraphics.
}
