% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodIndex.r
\name{periodIndex}
\alias{periodIndex}
\title{Calcualte the index according year and days at a specific period}
\usage{
periodIndex(year, day, period)
}
\arguments{
\item{year}{A vector of year}

\item{day}{A vector of year}

\item{period}{The periods to count}
}
\value{
Generated index according to period
}
\description{
Calcualte the index according year and days at a specific period
}
