% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSites.R
\docType{methods}
\name{plotSites,data.frame-method}
\alias{plotSites,data.frame-method}
\title{Plot site position in a map}
\usage{
\S4method{plotSites}{data.frame}(
  object,
  lines = ozRegion(),
  sitename = FALSE,
  col = "red",
  pch = 24,
  text.cex = 0.8,
  ...
)
}
\arguments{
\item{object}{A data.frame object.}

\item{lines}{A vector of lines for maps boundaries. Australia map from oz Package will be showed.}

\item{sitename}{Whether to show site name.}

\item{col}{The colour of site symbol.}

\item{pch}{The symbol type of site.}
}
\description{
Plot site position in a map
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
plotSites( records )
}
