% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSites.R
\docType{methods}
\name{plotSites,WeaAna-method}
\alias{plotSites,WeaAna-method}
\title{Plot site position in a map}
\usage{
\S4method{plotSites}{WeaAna}(
  object,
  lines = ozRegion(),
  sitename = FALSE,
  col = "red",
  pch = 24,
  ...
)
}
\arguments{
\item{object}{A WeaAna object.}

\item{lines}{A vector of lines for maps boundaries. Australia map from oz Package will be showed.}

\item{sitename}{Whether to show site name.}

\item{col}{The colour of site symbol.}

\item{pch}{The symbol type of site.}
}
\description{
Plot site position in a map
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
plotSites( records )
}
