% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSites.R
\name{plotSites}
\alias{plotSites}
\title{Plot site position in a map}
\usage{
plotSites(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Plot site position in a map
}
