% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWeatherRecords.r
\name{newPointer}
\alias{newPointer}
\alias{readWeatherRecords}
\alias{readSite}
\alias{readSiteAPSIM}
\alias{readSiteRDATA}
\alias{readSiteGHCN}
\title{Create new pointer object}
\usage{
newPointer(value)

readWeatherRecords(dataFiles = NULL, dataFolders = NULL,
  dataFormat = "APSIM", dataWeather = NULL, load.later = FALSE, ...)

readSite(filename, dataFormat = "APSIM", ...)

readSiteAPSIM(filename)

readSiteRDATA(filename)

readSiteGHCN(filename, ...)
}
\arguments{
\item{value}{The value for this pointer}

\item{dataFiles}{A character vector to specify the path of weather data files.}

\item{dataFolders}{A character vector to specify the path of weather data folders.}

\item{dataFormat}{The format of weather data file.}

\item{dataWeather}{A data.frame for existing data.}

\item{load.later}{Whether load weather records now or later. 
"dataFroamt" should be One of "APSIM" and "RDATA".}

\item{...}{Other arguments}

\item{filename}{The file name of weather data file.}

\item{dataFormat}{The format of weather data file.}

\item{...}{Other arguments}

\item{filename}{The file name of weather data file.}

\item{filename}{The file name of weather data file.}

\item{filename}{The file name of weather data file.}

\item{...}{Other arguments for site information}
}
\value{
A new pointer to value

A WeaAna class which contains all weather data.

A WeaAnaSite class which contains all weather data.

A WeaAnaSite class which contains all weather data.

A WeaAnaSite class which contains all weather data.

A WeaAnaSite class which contains all weather data.
}
\description{
Create new pointer object

Read weather records from a file list and/or a folder list

Read weather records from a weather data file

Read weather records from a weather data file with APSIM format

Read weather records from a weather data file with RDATA format

Read weather records from a weather data file with GHCN format
}
