% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerRes.R
\docType{methods}
\name{registerRes,WeaAna,character,character-method}
\alias{registerRes,WeaAna,character,character-method}
\title{Register name and type of new result}
\usage{
\S4method{registerRes}{WeaAna,character,character}(object, name, type)
}
\arguments{
\item{object}{A WeaAna object.}

\item{name}{Name of result}

\item{type}{Type of result}
}
\description{
Register name and type of new result
}
