% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerRes.R
\name{registerRes}
\alias{registerRes}
\title{Register name and type of new result}
\usage{
registerRes(object, name, type)
}
\arguments{
\item{object}{A WeaAna object.}

\item{name}{Name of result}

\item{type}{Type of result}
}
\description{
Register name and type of new result
}
