% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.weaanaClass.R
\docType{class}
\name{result-class}
\alias{result-class}
\title{Define the class for statistics results}
\description{
Define the class for statistics results
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of result}

\item{\code{type}}{Type of result}
}}

