% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{resultsInfor,WeaAna-method}
\alias{resultsInfor,WeaAna-method}
\title{Get name and type of results}
\usage{
\S4method{resultsInfor}{WeaAna}(object)
}
\arguments{
\item{object}{A WeaAna object.}
}
\description{
Get name and type of results
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
resultsInfor( records )

}
