% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{resultsInfor}
\alias{resultsInfor}
\title{Get name and type of results}
\usage{
resultsInfor(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Get name and type of results
}
