% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWeaAna.R
\docType{methods}
\name{show,WeaAna-method}
\alias{show,WeaAna-method}
\title{Show basic information of class WeaAna}
\usage{
\S4method{show}{WeaAna}(object)
}
\arguments{
\item{object}{WeaAna objects}
}
\description{
Show the name, number, latitude, longitude of all weather stations.
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
show( records )
records
}
