% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{siteInfor}
\alias{siteInfor}
\alias{siteInfor,WeaAna-method}
\alias{siteInfor,WeaAna,WeaAna-method}
\alias{siteInfor,WeaAnaSite-method}
\alias{siteInfor,WeaAnaSite,WeaAnaSite-method}
\title{Get site information}
\usage{
siteInfor(object, ...)

\S4method{siteInfor}{WeaAna}(object, load.now = FALSE)

\S4method{siteInfor}{WeaAnaSite}(object, load.now = FALSE)
}
\arguments{
\item{object}{A WeaAnaSite object.}

\item{...}{Not used}

\item{load.now}{Whether load site information}
}
\value{
Site information in the WeaAna object

Site information in the WeaAnaSite object
}
\description{
Get site information

Get site information

Get site information
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
siteInfor( records )
siteInfor( records, load.now = TRUE )
}
