% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{siteInfor,WeaAnaSite-method}
\alias{siteInfor,WeaAnaSite-method}
\title{Get site information}
\usage{
\S4method{siteInfor}{WeaAnaSite}(object, load.now = FALSE)
}
\arguments{
\item{object}{A WeaAnaSite object.}

\item{load.now}{Whether load site information}
}
\description{
Get site information
}
