% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileList.R
\name{siteList}
\alias{siteList}
\title{Get all file name from a vector of files}
\usage{
siteList(fileLists = NULL)
}
\arguments{
\item{fileLists}{A vector of files}
}
\value{
A vector of file name
}
\description{
Get all file name from a vector of files
}
