% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\docType{methods}
\name{siteNum,WeaAna-method}
\alias{siteNum,WeaAna-method}
\title{Site number}
\usage{
\S4method{siteNum}{WeaAna}(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{marker arguments}
}
\description{
Site number
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
siteNum( records )
}
