% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites2KML.R
\docType{methods}
\name{sites2KML,data.frame-method}
\alias{sites2KML,data.frame-method}
\title{Convert site position to KML file}
\usage{
\S4method{sites2KML}{data.frame}(object, file = "site.kml", name = TRUE, symbol = 1)
}
\arguments{
\item{object}{A data.frame object.}

\item{file}{The output file name.}

\item{name}{whether to show site name.}

\item{symbos}{The symbol for Google earth.}
}
\description{
Convert site position to KML file
}
\examples{
library( weaana )
data( "WeatherRecordsDemo" ) 
sites2KML( records )
}
