% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites2KML.R
\docType{methods}
\name{sites2KML,missing-method}
\alias{sites2KML,missing-method}
\title{Show all Google Earth symbols}
\usage{
\S4method{sites2KML}{missing}(object)
}
\arguments{
\item{object}{A data.frame object.}

\item{file}{The output file name.}

\item{whether}{to show site name.}
}
\description{
Show all Google Earth symbols
}
\examples{
sites2KML()
}
