% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites2KML.R
\name{sites2KML}
\alias{sites2KML}
\title{Convert site position to KML file}
\usage{
sites2KML(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Convert site position to KML file
}
