% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spatial}
\alias{spatial}
\title{Calculate the spatial slope and aspect}
\usage{
spatial(x, slope = TRUE, aspect = TRUE)
}
\arguments{
\item{x}{A matrix for spatial data with row for longitude and column for latitude.
dimnames must be specified for values of longitude and latitude}

\item{slope}{Logical, whether return slope}

\item{aspect}{Logical, whether return aspect}
}
\description{
Burrough, P. A., and McDonell, R. A., 1998. Principles of Geographical Information Systems (Oxford University Press, New York)
}
