% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericalDistance.R
\name{sphericalDistance}
\alias{sphericalDistance}
\title{Calculate the sphere distance}
\usage{
sphericalDistance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude}

\item{lon1}{Longitude}

\item{lat2}{Latitude}

\item{lon2}{Longitude}
}
\value{
Distance in km
}
\description{
Calculate the sphere distance
}
