% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{len}
\alias{len}
\alias{left}
\alias{right}
\alias{omitBlank}
\alias{searchChar}
\alias{omitBlankSE}
\alias{vector2string}
\title{String function in this package}
\usage{
len(str)

left(str, num = 1)

right(str, num = 1)

omitBlank(str)

searchChar(str, search, start = 1)

omitBlankSE(str)

vector2string(vector, sep = ", ")
}
\arguments{
\item{str}{Input string}

\item{num}{The number of character will be returned from right}

\item{search}{The character will be searched}

\item{start}{The start position to search}

\item{vector}{The input vector to convert}

\item{sep}{The separater with default value ", "}
}
\value{
The character number of the string

the character vector of the string from left

the character vector of the string from right

the character vector except the blank

The postion where character firstly appear

the character vector except the blank at the start and end

The string which connected all members of vector
}
\description{
Get the length of a string

Get the several characters of a string from left

Get the several characters of a string from right

Omit the blank of a string

Search a character from a string

Omit the start and end blank charcter

Convert vertor to string with a separater
}
