% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalTime.R
\docType{methods}
\name{thermalTime,WeaAna-method}
\alias{thermalTime,WeaAna-method}
\title{Calculate thermal time}
\usage{
\S4method{thermalTime}{WeaAna}(object, x_temp, y_temp, method = NULL)
}
\arguments{
\item{object}{WeaAna object}

\item{x_temp}{The cardinal temperatures}

\item{y_temp}{The effective thermal time}

\item{method}{The method to calculate thermal time. 
The default method is ( maxt + mint ) / 2 - base. 
The three hour temperature methods will be used if method = '3hr'}
}
\description{
Calculate thermal time
}
