% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalTime.R
\name{thermalTime}
\alias{thermalTime}
\title{Calculate thermal time}
\usage{
thermalTime(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate thermal time
}
