% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ttest_ts}
\alias{ttest_ts}
\title{Significantly t-test with autocorrelation for time serial data}
\usage{
ttest_ts(y, slope = NULL)
}
\arguments{
\item{x}{A vector of time serial data}
}
\description{
Method is presented by Santer et al. 2000
}
