% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd.apsim.r
\name{vpd.apsim}
\alias{vpd.apsim}
\title{Calculate the vpd using apsim method}
\usage{
vpd.apsim(maxt, mint)
}
\arguments{
\item{maxt}{The maximum temperature}

\item{mint}{The minimum temperature}
}
\value{
vpd
}
\description{
Calculate the vpd using apsim method
}
