% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{waGetPara}
\alias{waGetPara}
\title{Get parameters}
\usage{
waGetPara(name = NULL)
}
\arguments{
\item{name}{Variable name which will return. All parameters will return if NULL}
}
\description{
Get parameters
}
\examples{
library(weaana)
data( "WeatherRecordsDemo" ) 
waGetPara( "yrange" )
# for lower level parameters
waGetPara( "extreme$maxt$hot.day$value" )
}
