% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapply.R
\name{wapply}
\alias{wapply}
\title{Statistics weather records by a certain period and function}
\usage{
wapply(object, ...)
}
\arguments{
\item{object}{A WeaAna objects.}

\item{...}{Not used}
}
\description{
Statistics weather records by a certain period and function
}
