% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcal.R
\name{wcal}
\alias{wcal}
\title{Calculate weather variables through function or a string formula.}
\usage{
wcal(object, ...)
}
\arguments{
\item{object}{A WeaAna objects.}

\item{...}{Not used}
}
\description{
Calculate weather variables through function or a string formula.
}
