% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcal.R
\name{wcalFun}
\alias{wcalFun}
\title{Calculate weather variables through a function.}
\usage{
wcalFun(x, FUN, var.args, other.args = NULL, len = length(x[[1]]))
}
\arguments{
\item{x}{A data frame contained all weather records}

\item{FUN}{A function to be used which results should have the same length as original records.}

\item{var.args}{Arguments of weather variable pass to \code{FUN}.}

\item{other.args}{Optional arguments to \code{FUN}}

\item{len}{The length of result}
}
\description{
Calculate weather variables through a function.
}
