% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcal.R
\name{wcalStr}
\alias{wcalStr}
\title{Calculate weather variables through a string formula.}
\usage{
wcalStr(x, str = NULL, len = length(x[[1]]))
}
\arguments{
\item{x}{A data frame contained all weather records}

\item{str}{A string function}

\item{len}{The length of result}
}
\description{
Calculate weather variables through a string formula.
}
