% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.weaanaClass.R
\docType{class}
\name{WeaAna-class}
\alias{WeaAna-class}
\title{Define the class for multiple sites}
\description{
Define the class for multiple sites
}
\section{Slots}{

\describe{
\item{\code{num}}{total number of weather station}

\item{\code{records}}{A pointer vector to weather records of each site}

\item{\code{result}}{A pointer for all results name and type.}
}}

