% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.weaanaClass.R
\docType{class}
\name{WeaAnaSite-class}
\alias{WeaAnaSite-class}
\title{Define the class of WeaAna}
\description{
Define the class of WeaAna
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of weather station}

\item{\code{number}}{Station number of weather station}

\item{\code{latitude}}{Latitude of weather station}

\item{\code{longitude}}{Latitude of weather station}

\item{\code{tav}}{Annual average ambient temperature}

\item{\code{amp}}{Annual amplitude in mean monthly temperature}

\item{\code{marker}}{The extra marker for this site}

\item{\code{year}}{A vector of year of weather station}

\item{\code{day}}{A vector of day of weather station}

\item{\code{radn}}{A vector of radiation of weather station}

\item{\code{maxt}}{A vector of maximum temperature of weather station}

\item{\code{mint}}{A vector of minimum temperature of weather station}

\item{\code{evap}}{A vector of evaporation of weather station}

\item{\code{rain}}{A vector of rainfall of weather station}

\item{\code{vp}}{A vector of pressure atmosphere of weather station}

\item{\code{code}}{The 6 digit code indicates the source of the 6 data columns}

\item{\code{extra}}{A list of variables need to store}

\item{\code{res}}{All statistics results store in this slot}

\item{\code{figures}}{A list to store all plotted figures.}

\item{\code{file.path}}{The file path for this site.}

\item{\code{data.format}}{The data format for this site.}

\item{\code{load.later}}{Whether are records loaded laterly.}
}}

