% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeatherRecords.R
\docType{methods}
\name{writeWeatherRecords}
\alias{writeWeatherRecords}
\alias{writeWeatherRecords,WeaAna-method}
\alias{writeWeatherRecords,WeaAna,WeaAna-method}
\title{Write weather records into file}
\usage{
writeWeatherRecords(object, ...)

\S4method{writeWeatherRecords}{WeaAna}(object, file, cols = NULL)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}

\item{file}{Path of output file.}

\item{cols}{Columns to export. All columns exported if NULL}
}
\value{
No return values
}
\description{
Write weather records into file

Write weather records into file
}
