% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeatherRecords.R
\docType{methods}
\name{writeWeatherRecords,WeaAna-method}
\alias{writeWeatherRecords,WeaAna-method}
\title{Write weather records into file}
\usage{
\S4method{writeWeatherRecords}{WeaAna}(object, file)
}
\arguments{
\item{object}{A WeaAna object.}

\item{file}{Path of output file.}
}
\description{
Write weather records into file
}
