% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeWeatherRecords.R
\name{writeWeatherRecords}
\alias{writeWeatherRecords}
\title{Write weather records into file}
\usage{
writeWeatherRecords(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Write weather records into file
}
