% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyThermalTime.R
\docType{methods}
\name{yearlyThermalTime,WeaAna-method}
\alias{yearlyThermalTime,WeaAna-method}
\title{Calculate yearly thermal time}
\usage{
\S4method{yearlyThermalTime}{WeaAna}(object, yrange = waGetPara(object,
  "yrange"), base.temperature = waGetPara("base.temperature"),
  day.begin = waGetPara("day.begin"), day.end = waGetPara("day.end"))
}
\arguments{
\item{object}{A WeaAna object.}

\item{yrange}{The year range for statistics}

\item{base.temperature}{The base temperature to calculate the degree days}

\item{day.end}{The end day to calculate thermal time}

\item{day.start}{The start day to calculate thermal time}
}
\description{
Calculate yearly thermal time
}
