% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyThermalTime.R
\name{yearlyThermalTime}
\alias{yearlyThermalTime}
\title{Calculate yearly thermal time}
\usage{
yearlyThermalTime(object, ...)
}
\arguments{
\item{object}{A WeaAna object.}

\item{...}{Not used}
}
\description{
Calculate yearly thermal time
}
