% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWeatherRecords.r
\name{readWeatherRecords}
\alias{readWeatherRecords}
\title{Read weather records from a file list and/or a folder list}
\usage{
readWeatherRecords(
  dataFiles = NULL,
  dataFolders = NULL,
  dataFormat = "APSIM",
  dataWeather = NULL,
  load.later = FALSE,
  ...
)
}
\arguments{
\item{dataFiles}{A character vector to specify the path of weather data files.}

\item{dataFolders}{A character vector to specify the path of weather data folders.}

\item{dataFormat}{The format of weather data file.}

\item{dataWeather}{A data.frame for existing data.}

\item{load.later}{Whether load weather records now or later. 
"dataFroamt" should be One of "APSIM" and "RDATA".}

\item{...}{Other arguments}
}
\value{
A WeaAna class which contains all weather data.
}
\description{
Read weather records from a file list and/or a folder list
}
